Update HomePage.java
In HomePage I fixed the textField labels to match the right field. I then added a simple EventHandler to take the typed data and create a new account using the new Account class. This will store the data in Accounts.csv. We still need to make sure the input is formatted correctly and all that fun stuff.